TZ patch refactored
