+++ feincms/compat.py
@@ -1,3 +1,10 @@
+try:
+    from django.utils.timezone import now as compatible_now
+except ImportError:
+    # Django < 1.4 comes without timezone support
+    from datetime.datetime import now as compatible_now
+
+
 def c_any(iterable):
     """
     Implements python 2.5's any()
@@ -17,3 +24,5 @@ def c_all(iterable):
         if not element:
             return False
     return True
+
+
